# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rawmart Blog API",
        "description": "API documentation for the RAWMART BLOG ASSESSMENT API",
        "contact": {
            "email": "ibrahimsharafadeen95@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/profile": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Update user profile",
                "description": "Allows the authenticated user to update their profile information (name, email, avatar).",
                "operationId": "0b03b092c754a4ac5791587b08b0f285",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "j@doe.com"
                                    },
                                    "image": {
                                        "description": "User avatar image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/change-password": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Change user password",
                "description": "Allows the authenticated user to change their password by providing the current password and a new one.",
                "operationId": "9f45564026a03d891abdb6f387fa9798",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "oldpassword123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password changed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid current password"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "new-user"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "j@doe.com"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde"
                    },
                    "joined_at": {
                        "type": "string",
                        "example": "Joined Dec 2023"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-21T13:53:01.000000Z"
                    },
                    "articles": {
                        "type": "integer",
                        "example": 12
                    },
                    "comments": {
                        "type": "integer",
                        "example": 46
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}